fix write register data lost , group and simplify parameter

+++ component_driver/ppg/afe4404/afe4404_interface.h
@@ -10,7 +10,7 @@
 void hw_afe4404_event_handler(nrf_drv_twi_evt_t const * p_event, void * p_context);
 uint32_t hw_afe4404_init(void);
 int32_t hw_afe4404_register_read(uint8_t reg);
-uint32_t hw_afe4404_write_single_register(uint8_t reg, uint16_t data);
+uint32_t hw_afe4404_write_single_register(uint8_t reg, uint32_t data);
 void hw_afe4404_reset(void);
 
 #endif // AFE_INTERFACE_H
